framework/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
JayDeBeApi==1.2.3
JPype1==0.7.5
numpy==1.19.0
pytz==2019.3
pywin32==228
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
